# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Minibank API
  description: |
    A reactive banking system API built with Spring Boot 3.4.1, Kotlin, and R2DBC.
    
    ## Features
    - Client management with full CRUD operations
    - Bank account management with multiple account types
    - Secure money transfers between accounts
    - Transaction history and reporting
    - Real-time balance updates
    - Multi-currency support (USD, EUR, GBP)
    - Money-safe operations with 4-decimal precision
    
    ## Security
    - OAuth2/OIDC authentication
    - Role-based access control
    - JWT token-based authorization
    
    ## Money Safety
    All monetary operations use BigDecimal with 4-decimal precision and banker's rounding (HALF_EVEN).
    Transaction limits: 0.0001 to 999,999,999,999,999.9999 per currency.
  version: 0.5.0
  contact:
    name: EMeMSoft spol. s r.o.
    email: support@ememsoft.cz
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8082
    description: Development server
  - url: https://api.minibank.ememsoft.cz
    description: Production server
...
paths:
...
  /transactions/transfer:
    post:
      tags:
        - transactions
      summary: Transfer money between accounts
      description: |
        Performs a money transfer between two accounts with the following validations:
        - Both accounts must exist and have matching currencies
        - Source account must have sufficient funds
        - Amount must be between 0.0001 and 999,999,999,999,999.9999
        - Accounts must be different
      operationId: transferMoney
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '201':
          description: Transfer completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResponse'
        '400':
          description: Invalid transfer request
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/InsufficientFundsError'
                  - $ref: '#/components/schemas/CurrencyMismatchError'
                  - $ref: '#/components/schemas/InvalidAmountError'
                  - $ref: '#/components/schemas/SameAccountError'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - BearerAuth: [write:transactions]
  /transactions/{id}:
    get:
      tags:
        - transactions
      summary: Get transaction by ID
      description: Retrieves a specific transaction by its unique identifier
      operationId: getTransactionById
      parameters:
        - name: id
          in: path
          required: true
          description: Transaction ID
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: Transaction found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResponse'
        '404':
          $ref: '#/components/responses/TransactionNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - BearerAuth: [read:transactions]
  /transactions/account/{accountId}:
    get:
      tags:
        - transactions
      summary: Get all transactions for an account
      description: Retrieves all transactions (incoming and outgoing) for a specific account
      operationId: getTransactionsByAccount
      parameters:
        - name: accountId
          in: path
          required: true
          description: Account ID
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: Account transactions retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransferResponse'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - BearerAuth: [read:transactions]
  /transactions/account/{accountId}/outgoing:
    get:
      tags:
        - transactions
      summary: Get outgoing transactions for an account
      description: Retrieves all outgoing transactions from a specific account
      operationId: getOutgoingTransactions
      parameters:
        - name: accountId
          in: path
          required: true
          description: Account ID
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: Outgoing transactions retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransferResponse'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - BearerAuth: [read:transactions]
  /transactions/account/{accountId}/incoming:
    get:
      tags:
        - transactions
      summary: Get incoming transactions for an account
      description: Retrieves all incoming transactions to a specific account
      operationId: getIncomingTransactions
      parameters:
        - name: accountId
          in: path
          required: true
          description: Account ID
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: Incoming transactions retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransferResponse'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - BearerAuth: [read:transactions]
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from Keycloak authentication
  schemas:
...
    TransferRequest:
      type: object
      required:
        - sourceAccountId
        - targetAccountId
        - amount
        - currency
      properties:
        sourceAccountId:
          type: integer
          format: int64
          description: ID of the account to transfer from
          example: 1
        targetAccountId:
          type: integer
          format: int64
          description: ID of the account to transfer to
          example: 2
        amount:
          type: string
          format: decimal
          description: Amount to transfer (0.0001 to 999,999,999,999,999.9999)
          example: "150.0000"
          pattern: '^\d{1,15}\.\d{1,4}$'
        currency:
          $ref: '#/components/schemas/Currency'
        reference:
          type: string
          description: Optional transfer reference or description
          example: "Monthly payment"
          maxLength: 255
    TransferResponse:
      type: object
      required:
        - id
        - sourceAccountId
        - targetAccountId
        - amount
        - currency
        - timestamp
        - status
        - reference
      properties:
        id:
          type: integer
          format: int64
          description: Unique transaction identifier
          example: 1
        sourceAccountId:
          type: integer
          format: int64
          description: ID of the source account
          example: 1
        targetAccountId:
          type: integer
          format: int64
          description: ID of the target account
          example: 2
        amount:
          type: string
          format: decimal
          description: Transfer amount with 4 decimal precision
          example: "150.0000"
        currency:
          $ref: '#/components/schemas/Currency'
        timestamp:
          type: string
          format: date-time
          description: When the transaction occurred
          example: "2025-01-15T10:30:00Z"
        status:
          $ref: '#/components/schemas/TransactionStatus'
        reference:
          type: string
          description: Transfer reference or description
          example: "Monthly payment"
...
    Currency:
      type: string
      enum:
        - USD
        - EUR
        - GBP
      description: ISO 4217 currency code
      example: "USD"
    TransactionStatus:
      type: string
      enum:
        - PENDING
        - COMPLETED
        - FAILED
      description: Current status of the transaction
      example: "COMPLETED"
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type identifier
          example: "Client Not Found"
        message:
          type: string
          description: Human-readable error description
          example: "Client with ID 123 not found"
    InsufficientFundsError:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
      example:
        error: "Insufficient Funds"
        message: "Insufficient funds in account 1. Available: 100.00, Required: 150.00"
    CurrencyMismatchError:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
      example:
        error: "Currency Mismatch"
        message: "Currency mismatch: source account has USD, but transfer requests EUR"
    InvalidAmountError:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
      example:
        error: "Invalid Transaction Amount"
        message: "Transaction amount must be positive"
    SameAccountError:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
      example:
        error: "Invalid Transfer"
        message: "Cannot transfer money to the same account"
  responses:
...
    AccountNotFound:
      description: Account not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Account Not Found"
            message: "Account with ID 456 not found"
    TransactionNotFound:
      description: Transaction not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Transaction Not Found"
            message: "Transaction with ID 789 not found"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Internal Server Error"
            message: "An unexpected error occurred"
...
security:
  - BearerAuth: []

# Additional API Documentation
...